restructured userContest by 1.moving useEffect( 'to initiate google sign up button' ) to sign-up component 2.changed where currentUser and setCurrentUser live by moving the state to app.js from userContext 3. and wrap the entire routes with UserContext provide and pass in value for children component to use it 4.deleted sign-up-google component because it is combined in the sign-up component
