+++ client/src/App.js
@@ -1,22 +1,27 @@
 import SignUp from "../src/components/sign-up/sign-up.component.js";
 import SignIn from "../src/components/sign-in/sign-in.component.js";
-import SignUpGoogle from "../src/components/sign-up-google/sign-up-google.component.js";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import MainPage from "./components/main-page/main-page.component.js";
 import HomePage from "./components/home-page/home-page.component";
 import CreateRecipe from "./components/create-recipe/create-recipe.component.js";
+import { UserContext } from "./contexts/user.context";
+import { useState } from "react";
 
 const App = () => {
+  const [currentUser, setCurrentUser] = useState({});
+  const value = { currentUser, setCurrentUser };
+
   return (
     <Router>
-      <Routes>
-        <Route path="/" element={<MainPage />} />
-        <Route path="/signup" element={<SignUp />} />
-        <Route path="/signin" element={<SignIn />} />
-        <Route path="/signupgoogle" element={<SignUpGoogle />} />
-        <Route path="/home" element={<HomePage />} />
-        <Route path="/create-recipe" element={<CreateRecipe />} />
-      </Routes>
+      <UserContext.Provider value={value}>
+        <Routes>
+          <Route path="/" element={<MainPage />} />
+          <Route path="/signup" element={<SignUp />} />
+          <Route path="/signin" element={<SignIn />} />
+          <Route path="/home" element={<HomePage />} />
+          <Route path="/create-recipe" element={<CreateRecipe />} />
+        </Routes>
+      </UserContext.Provider>
     </Router>
   );
 };
