+++ client/src/index.js
@@ -1,16 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { UserProvider } from "../src/contexts/user.context";
-// import './index.css';
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <UserProvider>
-      <App />
-    </UserProvider>
+    <App />
   </React.StrictMode>
 );
